refactor(ui): extract api/runtime.ts + stats.ts + memories.ts + config.ts, barrel the rest
Finish the api.ts decomposition. The remaining endpoints split into
four focused domain modules: runtime (readiness probe, runtime info,
project list), stats (stats/usage/session/raw-event loaders),
memories (memory + summary pagination, visibility toggle, forget,
pack trace), and config (observer status, config load/save). api.ts
becomes a thin barrel re-exporting every public function + type so
existing `import * as api from "../lib/api"` call sites stay on the
same path.
